solución de un error en la carga de imagenes en el home

+++ android_app/app/src/main/java/com/example/aplicacionmovil/ui/main/BuyButton.kt
@@ -133,7 +133,7 @@ fun BuyButton(
             onConfirm = {
                 viewModel.createPurchase(
                     articuloId = article.id,
-                    precio = article.precio.toDouble()
+                    precio = article.precio
                 )
             },
             onDismiss = {
